+++ internal/rest/server.go
@@ -77,9 +77,9 @@ func NewServer(cfg config.REST, authStorage AuthStorage, cl *coresdk.Client, sc 
 	handler.HandleFunc("/feed/mark-as-read", srv.markAsReadBatch).Methods(http.MethodPost).Name("mark_as_read_batch")
 	handler.HandleFunc("/feed/{id}/mark-as-read", srv.markFeedItemAsRead).Methods(http.MethodPost).Name("mark_feed_item_as_read")
 
-	handler.HandleFunc("/settings/notifications", srv.storePushToken).Methods(http.MethodPost).Name("store_push_token")
-	handler.HandleFunc("/settings/notifications", srv.tokenExists).Methods(http.MethodGet).Name("push_token_exists")
-	handler.HandleFunc("/settings/notifications", srv.removePushToken).Methods(http.MethodDelete).Name("remove_push_token")
+	handler.HandleFunc("/notifications/settings", srv.storePushToken).Methods(http.MethodPost).Name("store_push_token")
+	handler.HandleFunc("/notifications/settings", srv.tokenExists).Methods(http.MethodGet).Name("push_token_exists")
+	handler.HandleFunc("/notifications/settings", srv.removePushToken).Methods(http.MethodDelete).Name("remove_push_token")
 
 	return srv
 }